refactor: clean up some more
