+++ components/index.js
@@ -11,7 +11,7 @@ import FullWidthImageSmall from './full-width-image-small/full-width-image-small
 import InlineImage from './inline-image/inline-image'
 import Layout from './layout/layout'
 import LogoCarousel from './logo-carousel/logo-carousel'
-import MailchimpForm from './mailchimp-form/mailchimp-form'
+import MailchimpForm from './mailchimp/mailchimp-form'
 import MenuBar from './menu-bar/menu-bar'
 import RichBodyText from './rich-body-text/rich-body-text'
 import SocialShare from './social-share/social-share'
