+++ components/mailchimp/mailchimp.config.js
@@ -1,0 +1,8 @@
+export default {
+	'company-name': { id: 'MMERGE6', type: 'text' },
+	'email': { id: 'EMAIL', type: 'email' },
+	'first-name': { id: 'FNAME', type: 'text' },
+	'job-title': { id: 'MMERGE5', type: 'text' },
+	'last-name': { id: 'LNAME', type: 'text' },
+	'phone-number': { id: 'PHONE', type: 'phone' },
+}
